+++ baseapp_auth/rest_framework/pre_auth/serializers.py
@@ -9,7 +9,7 @@ from baseapp_auth.tokens import PreAuthTokenGenerator
 User = get_user_model()
 
 
-class AuthTokenPreAuthSerializer(serializers.Serializer):
+class BasePreAuthSerializer(serializers.Serializer):
     token = serializers.CharField()
 
     def validate_token(self, token):
@@ -25,25 +25,12 @@ class AuthTokenPreAuthSerializer(serializers.Serializer):
             raise serializers.ValidationError(_("Invalid token."))
         return token
 
+
+class AuthTokenPreAuthSerializer(BasePreAuthSerializer):
     def save(self):
         return Token.objects.get_or_create(user=self.user)[0]
 
 
-class JWTPreAuthSerializer(serializers.Serializer):
-    token = serializers.CharField()
-
-    def validate_token(self, token):
-        generator = PreAuthTokenGenerator()
-        value = generator.decode_token(token)
-        if value is None:
-            raise serializers.ValidationError(_("Invalid token."))
-        try:
-            self.user = User.objects.get(id=value[0])
-        except User.DoesNotExist:
-            raise serializers.ValidationError(_("Invalid token."))
-        if not generator.is_value_valid(self.user, value):
-            raise serializers.ValidationError(_("Invalid token."))
-        return token
-
+class JWTPreAuthSerializer(BasePreAuthSerializer):
     def save(self):
         return RefreshToken.for_user(self.user)
